Commit de Finalizacion microservicio Votaciones

+++ Src/Services/Votaciones/Votaciones.Domain/Votacion.cs
@@ -7,9 +7,9 @@ namespace Votaciones.Domain
     {
         [Key]
         public long Id { get; set; }
-        public int Id_Eleccion { get; set; }
-        public int Id_Candidato { get; set; }
-        public int Id_Sufragante { get; set; }
+        public long Id_Eleccion { get; set; }
+        public long Id_Candidato { get; set; }
+        public long Id_Sufragante { get; set; }
         public int Salto { get; set; }
     }
 }
