+++ Src/Services/Votaciones/Votaciones.Service.EventHandlers/Command/VotacionCreateCommand.cs
@@ -7,9 +7,9 @@ namespace Votaciones.Service.EventHandlers.Command
 {
     public class VotacionCreateCommand : INotification
     {
-        public int Id_Eleccion { get; set; }
-        public int Id_Candidato { get; set; }
-        public int Id_Sufragante { get; set; }
+        public long Id_Eleccion { get; set; }
+        public long Id_Candidato { get; set; }
+        public long Id_Sufragante { get; set; }
         public int Salto { get; set; }
     }
 }
