+++ Src/Services/Votaciones/Votaciones.Service.Query/DTOs/VotacionDto.cs
@@ -7,9 +7,9 @@ namespace Elecciones.Service.Query.DTOs
     public class VotacionDto
     {
         public long Id { get; set; }
-        public int Id_Eleccion { get; set; }
-        public int Id_Candidato { get; set; }
-        public int Id_Sufragante { get; set; }
+        public long Id_Eleccion { get; set; }
+        public long Id_Candidato { get; set; }
+        public long Id_Sufragante { get; set; }
         public int Salto { get; set; }
     }
 }
